feat: Implement Smart Auto-Zoom functionality and enhance frame padding controls
- Added Smart Auto-Zoom feature that automatically places focus regions based on cursor activity during video playback.
- Introduced new utility functions for calculating frame padding in pixels based on a percentage of the shorter video edge.
- Updated the EditorStore to manage auto-zoom states and added methods for adding and clearing auto-generated zoom regions.
- Enhanced the BackgroundPicker component to allow frame padding to be set as a percentage instead of fixed pixels.
- Refactored ZoomSuggestionsPopover to utilize new auto-zoom utilities and improved zoom region placement logic.
- Improved UI components (FocusOverlay, AnnotationOverlay, TextAnnotationLayer) to account for dynamic frame padding.
- Added validation to ensure only valid image values are accepted in the BackgroundPicker.
- Updated the custom title bar buttons for better accessibility and visual consistency.

+++ apps/desktop/src-tauri/src/render/node_types.rs
@@ -134,7 +134,8 @@ pub struct BackgroundNode {
     pub background_type: String,
     pub value: String,
     pub blur: f64,
-    pub padding: u32,
+    /// Frame padding as percent of the shorter source edge.
+    pub padding: f64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
